thing/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
psycopg2-binary==2.9.6
Pygments==2.15.1
PyJWT==2.6.0
PySocks==1.7.1
pytz==2023.3
requests==2.29.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
urllib3==1.26.15
whitenoise==6.4.0